% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mboost.R
\name{check_mboost}
\alias{check_mboost}
\title{Diagnostic function for mboost objects}
\usage{
check_mboost(object, plot = TRUE, what = c("inbagrisk", "edf1", "edf2",
  "AIC1", "AIC2", "AICc1", "AICc2", "BIC1", "BIC2", "gMDL1", "gMDL2",
  "selectionFreq", "cumulativeExplainedRisk"), FUN_iter = list(),
  FUN_obj = list(), ups = NULL, ...)
}
\arguments{
\item{object}{an mboost object}

\item{plot}{logical; whether to show diagnostic plots}

\item{what}{a character vector specifying what is computed per 
default. Additional measures can be defined by \code{FUN_iter} and 
\code{FUN_obj}.}

\item{FUN_iter}{a named list of functions
that are applied over the all iterations of the mboost object
and returned in the results}

\item{FUN_obj}{a named list of functions
that are applied on the given mboost object and return
in the results}

\item{ups}{if available a list of residual matrices of each iteration
as returned by the \code{getUpsilons} function.}

\item{...}{further paramters passed to ??}
}
\value{
A data.frame with diagnostic measures
}
\examples{
cars.gb <- gamboost(dist ~ speed, data = cars, dfbase = 4,
                    control = boost_control(mstop = 50))
res <- check_mboost(cars.gb)
str(res,1)
}
